requirements/dev.txt
-r base.txt
nox
pre-commit
mypy>=0.910
types-attrs
types-setuptools
pylint==2.12.2
astroid==2.9.3
pylint-pydantic
pyenchant
flake8
flake8-mypy-fork
flake8-docstrings
flake8-typing-imports
